backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
python-multipart==0.0.9
pdfplumber==0.11.0
openpyxl==3.1.2
reportlab==4.2.0
PyYAML==6.0.1
pytest==8.2.0
httpx==0.27.0
